requirements.txt
openai
requests
aiohttp
litellm
python-dotenv